requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
Django==2.2.4
djangorestframework==3.10.2
django-heroku==0.3.1
freezegun==0.3.12
gunicorn==19.9.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
